requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0
fastapi==0.115.14
google-auth==2.23.4
google-auth-oauthlib==1.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.0.2
openai==1.93.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
regex==2024.11.6
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
Werkzeug==3.1.3
zipp==3.23.0
